requirements.txt
pytest>=6.0.1
mypy>=0.782
topic-db>=1.0.0
python-slugify>=4.0.1